+++ src/python/DAS/analytics/analytics_web.py
@@ -3,7 +3,6 @@ import cherrypy
 import collections
 import time
 import json
-import atexit
 
 from DAS.analytics.analytics_config import DASAnalyticsConfig
 from DAS.web.das_webmanager import DASWebManager
@@ -12,6 +11,10 @@ from DAS.analytics.analytics_task import Task
 import DAS
 
 #This extensively borrows from DAS.web
+
+def json_requested():
+    return 'accept' in cherrypy.request.headers and \
+           'json' in cherrypy.request.headers['accept']
 
 DASAnalyticsConfig.add_option("web_port",
                               type=int,
@@ -25,7 +28,6 @@ DASAnalyticsConfig.add_option("web_base",
                               type=basestring,
                               default='/analytics',
       help="Base path for analytics web")
-
 class DASAnalyticsWebManager(DASWebManager):
     """
     Use the DASWebManager class as the root of our interfaces,
@@ -36,12 +38,40 @@ class DASAnalyticsWebManager(DASWebManager):
         self.root = root
         DASWebManager.__init__(self, config)
         self.base = config.web_base
+    
+    def templatepage(self, tmpl, **kwargs):
+        """
+        Intercept template requests, and return a raw dictionary if json.
+        """
+        if json_requested():
+            kwargs['template'] = tmpl
+            return json.dumps(kwargs, default=str)
+        else:
+            return super(DASAnalyticsWebManager, self).templatepage(tmpl, **kwargs)
         
+    def page(self, content):
+        """
+        Intercept page requests, and return no headers if json.
+        """
+        if json_requested():
+            return content
+        else:
+            return super(DASAnalyticsWebManager, self).page(content)
+    
     @cherrypy.expose
     def index(self, *args, **kwargs):
         "Show a main page with a list of other pages."
         return self.page(self.templatepage("analytics_main",
                                            base=self.base))
+    
+    def error(self, error, **kwargs):
+        page = self.templatepage("analytics_error",
+                                 error=error,
+                                 request=cherrypy.request.request_line,
+                                 params=cherrypy.request.params,
+                                 headers=cherrypy.request.headers,
+                                 extra=kwargs)
+        return self.page(page)
     
     def top(self):
         """
@@ -50,6 +80,12 @@ class DASAnalyticsWebManager(DASWebManager):
         return self.templatepage('analytics_header',
                                  base=self.base,
                                  yui=self.yuidir)
+
+    @cherrypy.expose
+    def doc(self):
+        """Return templated (essentially static) documentation page"""
+        return self.page(self.templatepage("analytics_doc",
+                                           base=self.base))
 
     def bottom(self):
         """
@@ -65,20 +101,21 @@ class DASAnalyticsWebManager(DASWebManager):
         task_schedule = self.root.controller.scheduler.get_scheduled()
         task_schedule = sorted(task_schedule, key=lambda x: x['at'])
         next = None
+        now = time.time()
         if 'next' in attrs:
             next = int(attrs['next'])
-            before = time.time() + next
+            before = now + next
             task_schedule = filter(lambda x: x['at'] < before, 
                                    task_schedule)
         
         task_running = self.root.controller.scheduler.get_running()
         
         task_running = sorted(task_running, key=lambda x: x['started'])
-            
         page = self.templatepage("analytics_schedule",
                                  schedule=task_schedule,
                                  running=task_running,
-                                 base=self.base)
+                                 base=self.base,
+                                 now=now)
         return self.page(page)
     
     @cherrypy.expose
@@ -148,15 +185,6 @@ class DASAnalyticsWebManager(DASWebManager):
                                  total=total,
                                  level=level)
         return self.page(page)
-            
-    
-    @cherrypy.expose
-    def shutdown(self, *path, **attrs):
-        """
-        Shut down the main loop and kill the cherrypy server.
-        Should probably require some auth, if kept at all
-        """
-        self.root.controller.stop()
     
     @cherrypy.expose
     def task(self, *path, **attrs):
@@ -164,14 +192,18 @@ class DASAnalyticsWebManager(DASWebManager):
         Show detailed information for a single path.
         """
         if not 'id' in attrs:
-            return self.page("No id specified. Cannot display.")
+            return self.error("Argument 'id' required")
         master_id = attrs['id']
         task = self.root.controller.scheduler.get_task(master_id)
-        if not task:
-            return self.page("Specified id does not exist. Cannot display")
-        page = self.templatepage('analytics_task',
-                                 base=self.base,
-                                 **task)
+                
+        if task:
+            page = self.templatepage('analytics_task',
+                                     base=self.base,
+                                     id=master_id,
+                                     **task)
+        else:
+            return self.error("Requested ID does not exist", 
+                              id=master_id)
         return self.page(page)
         
     
@@ -180,10 +212,24 @@ class DASAnalyticsWebManager(DASWebManager):
         """
         Show a page with some control options.
         """
+        
+        if 'key' in attrs and 'value' in attrs:
+            try:
+                value = json.loads(attrs['value'])
+            except:
+                return self.error("JSON decode of value failed",
+                                  value=attrs['value'])
+            result = self.config.set_option(attrs['key'], value)
+            if not result == True:
+                return self.error("Setting of option failed",
+                                  response=result)
+                
+        
         page = self.templatepage('analytics_control',
                                  base=self.base,
-                                 config=self.config._options)
+                                 config=self.config.get_dict())
         return self.page(page)
+        
     
     @cherrypy.expose
     def results(self, *path, **attrs):
@@ -253,7 +299,7 @@ class DASAnalyticsWebManager(DASWebManager):
         (eg requesting graphs, etc)
         """
         if not 'id' in attrs:
-            return self.page("No id specified. Cannot display.")
+            return self.error("Argument 'id' required")
         task_id = attrs['id']
         result = None
         for r in self.root.result_data:
@@ -261,7 +307,8 @@ class DASAnalyticsWebManager(DASWebManager):
                 result = r
                 break
         if not result:
-            return self.page("Specified id not found. Cannot display.")
+            return self.error("Requested ID does not exist", 
+                              id=task_id)
         
         page = self.templatepage('analytics_result',
                                  base=self.base,
@@ -275,13 +322,14 @@ class DASAnalyticsWebManager(DASWebManager):
         Remove a task from the schedule. TODO: Auth.
         """
         if not 'id' in attrs:
-            return self.page("No id specified. Cannot remove.")
+            return self.error("Argument 'id' required")
         master_id = attrs['id']
         result = self.root.controller.scheduler.remove_task(master_id)
         if result:
-            return self.page("Task removed.")
+            return self.schedule()
         else:
-            return self.page("Task removal failed. Probably didn't exist.")
+            return self.error("Unable to remove task. It probably didn't exist.",
+                              id=master_id)
     
     @cherrypy.expose
     def reschedule_task(self, *path, **attrs):
@@ -289,26 +337,29 @@ class DASAnalyticsWebManager(DASWebManager):
         Reschedule a task to a new time. TODO: Auth
         """
         if not ('id' in attrs and 'at' in attrs):
-            return self.page("Either (id,at) not specified. Cannot reschedule.")
+            return self.error("Arguments 'id' and 'at' required")
         master_id = attrs['id']
-        when = int(attrs['at'])
+        when = float(attrs['at'])
         result = self.root.controller.scheduler.reschedule_task(master_id, 
                                                                 when)
         if result:
-            return self.page("Task rescheduled.")
+            return self.schedule()
         else:
-            return self.page("Task rescheduling failed. Probably didn't exist.")
+            return self.error("Unable to reschedule task. It probably didn't exist.",
+                              id=master_id,
+                              at=when)
         
     @cherrypy.expose
     def add_task(self, *path, **attrs):
         """
         Add a new task, for an existing class.
         TODO: Auth
+        TODO: accept a single dictionary containing all arguments
         """
         if not ('name' in attrs 
                 and 'classname' in attrs 
                 and 'interval' in attrs):
-            return self.page("Require at least (name,classname,interval).")
+            return self.error("Arguments 'name', 'classname' and 'interval' required.")
         
         try:
             kwargs = json.loads(attrs.get('kwargs', "{}"))
@@ -318,20 +369,21 @@ class DASAnalyticsWebManager(DASWebManager):
                 kwargs[k.encode('ascii')] = v
             task = Task(name=attrs['name'],
                         classname=attrs['classname'],
-                        interval=int(attrs['interval']),
+                        interval=float(attrs['interval']),
                         kwargs=kwargs,
                         only_once=json.loads(attrs.get('only_once', 'false')),
                         max_runs=json.loads(attrs.get('max_runs', 'null')),
                         only_before=json.loads(attrs.get('only_before', 'null')))
-                        
+        except:
+            return self.error("There was an error decoding the arguments.")
+        try:
             self.root.controller.scheduler.add_task(task,
                                 when=json.loads(attrs.get('when', 'null')),
                                 offset=json.loads(attrs.get('offset', 'null')))
             
-            return self.page("Created new task with master_id=%s" % \
-                             task.master_id)
+            return self.task(id=task.master_id)
         except:
-            return self.page("There was an error creating your task.")
+            return self.error("There was an error adding the task.")
     
 
 class DASAnalyticsWeb:
@@ -351,7 +403,7 @@ class DASAnalyticsWeb:
         Set up the cherrypy server.
         """
         self.controller.logger.addHandler(WebHandler(self.log_data))
-        self.controller.scheduler.add_callback(self.result_data.append)
+        self.controller.scheduler.add_callback(self.result_data.appendleft)
         
         cherrypy.config["engine.autoreload_on"] = False
         cherrypy.config["server.socket_port"] = self.config.web_port